Workaround Chrome issues when using OpenSSL-based TLS.
This stores the client certificate chain in a HttpSession object
so if SSL session reuse occurs, we can still access the client chain.

The HTTP session is verified against the SSL session ID; you need both
to perform a replay attack; if you already have the SSL session ID, then
it is likely you already own the client.